# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Content Generation Service",
    "description": "AI-powered content generation service",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://content-generation.mcpfactory.org"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "PromptVariable": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Variable name as referenced in the prompt body via {{name}}."
          },
          "description": {
            "type": "string",
            "description": "Free-form description of what the caller should put for this variable. Caller decides the JSON shape — string, array, object, whatever fits the template. Multibrand is the default, so brand-related variables typically receive arrays or objects, not scalars."
          }
        },
        "required": [
          "name",
          "description"
        ]
      },
      "CreatePromptRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Unique prompt identifier, e.g. 'cold-email' or 'welcome-email'"
          },
          "prompt": {
            "type": "string",
            "description": "Prompt template text with {{variable}} placeholders. Must NOT contain company-specific data — only {{variables}}."
          },
          "variables": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PromptVariable"
            },
            "description": "Inputs the template expects. Each entry is { name, description }; the caller decides the JSON shape per name."
          }
        },
        "required": [
          "type",
          "prompt",
          "variables"
        ]
      },
      "VersionPromptRequest": {
        "type": "object",
        "properties": {
          "sourceType": {
            "type": "string",
            "description": "The type of the prompt to create a new version from, e.g. 'cold-email'"
          },
          "prompt": {
            "type": "string",
            "description": "New prompt template text with {{variable}} placeholders. Must NOT contain company-specific data — only {{variables}}."
          },
          "variables": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PromptVariable"
            },
            "description": "Inputs the template expects. Each entry is { name, description }; the caller decides the JSON shape per name."
          }
        },
        "required": [
          "sourceType",
          "prompt",
          "variables"
        ]
      },
      "PromptResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "prompt": {
            "type": "string"
          },
          "variables": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PromptVariable"
            }
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "prompt",
          "variables",
          "createdAt",
          "updatedAt"
        ]
      },
...
    },
...
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    }
  },
  "paths": {
...
    "/prompts": {
      "get": {
        "tags": [
          "Prompts"
        ],
        "summary": "Get a prompt template by type",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Prompt type to look up"
            },
            "required": true,
            "description": "Prompt type to look up",
            "name": "type",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Prompt found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing type query param",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Prompt not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Prompts"
        ],
        "summary": "Create a prompt template (idempotent — no-op if type already exists)",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePromptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prompt already exists (no-op)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptResponse"
                }
              }
            }
          },
          "201": {
            "description": "Prompt created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Prompts"
        ],
        "summary": "Create or version a prompt (auto-increments type name if sourceType exists)",
        "description": "If the prompt and variables are identical to the existing sourceType, returns the existing prompt as-is (200). If sourceType already exists but content differs, creates a new prompt with an auto-incremented type name (201). E.g. sourceType 'cold-email' → creates 'cold-email-v2'. sourceType 'cold-email-v5' → creates 'cold-email-v6'. If sourceType does not exist, creates the prompt with that type directly (201). The source prompt is never modified.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionPromptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prompt unchanged — returned existing version as-is",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptResponse"
                }
              }
            }
          },
          "201": {
            "description": "New versioned prompt created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/platform-prompts": {
      "get": {
        "tags": [
          "Prompts"
        ],
        "summary": "Get a prompt template by type (no identity headers required)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Prompt type to look up"
            },
            "required": true,
            "description": "Prompt type to look up",
            "name": "type",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Prompt found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing type query param",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Prompt not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Prompts"
        ],
        "summary": "Create a prompt template (idempotent — no-op if type already exists, no identity headers required)",
        "description": "Used at cold start to register prompt templates. Same pattern as key-service POST /platform-keys.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePromptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prompt already exists (no-op)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptResponse"
                }
              }
            }
          },
          "201": {
            "description": "Prompt created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}